+++ capycli/bom/findsources.py
@@ -19,7 +19,7 @@ from bs4 import BeautifulSoup
 from colorama import Fore, Style
 
 # from packageurl import PackageURL
-from cyclonedx.model import ExternalReferenceType
+from cyclonedx.model import ExternalReferenceType, XsUri
 from cyclonedx.model.bom import Bom
 from cyclonedx.model.component import Component
 from sw360 import SW360Error
@@ -469,7 +469,11 @@ class FindSources(capycli.common.script_base.ScriptBase):
     @staticmethod
     def find_source_url_by_language(component: Component) -> str:
         capycli.dependencies.javascript.GetJavascriptDependencies().try_find_component_metadata(component, "")
-        return CycloneDxSupport.get_ext_ref_source_url(component)
+        url = CycloneDxSupport.get_ext_ref_source_url(component)
+        if isinstance(url, XsUri):
+            return url._uri
+        else:
+            return url
 
     def find_sources(self, bom: Bom) -> Tuple[int, int]:
         """Go through the list of SBOM items and try to determine the source code."""