+++ capycli/common/capycli_bom_support.py
@@ -145,7 +145,10 @@ class CycloneDxSupport():
                 break
 
         if ext_ref:
-            ext_ref.url = XsUri(value)
+            if isinstance(value, XsUri):
+                ext_ref.url = value
+            else:
+                ext_ref.url = XsUri(value)
         else:
             CycloneDxSupport.set_ext_ref(comp, type, comment, value)